:recycle: renames rdpc proxy to file storage api
