+++ src/routes/file-storage-api/handlers/entitiesIdHandler.ts
